Move validateProjectDirectory to validators helper
